DXE-43 Add support for visitor_prioritization (VP) cloudlet
